+++ packages/common/web/src/types/biCallbacksTypes.ts
@@ -1,6 +1,20 @@
+import { ToolbarType } from './toolbarEnums';
+interface biCallbackParams {
+  version?: string;
+}
+
+type EntryType = ToolbarType;
+export interface onPluginAddStepArgs extends biCallbackParams {
+  pluginId: string;
+  pluginDetails: unknown;
+  entryPoint: ToolbarType;
+  entryType: EntryType;
+  step: 'FileUploadDialog' | 'PluginModal';
+}
 export interface BICallbacks {
   onPluginAdd?(pluginId: string, entryPoint: string, version: string): void;
   onPluginAddSuccess?(pluginId: string, entryPoint: string, version: string): void;
+  onPluginAddStep?(params: onPluginAddStepArgs): void;
   onPluginDelete?(pluginId: string, version: string): void;
   onPublish?(
     postId: string | undefined,