+++ packages/toolbars/web/src/Toolbar/ToolbarContainer.jsx
@@ -1,3 +1,4 @@
+/* eslint-disable jsx-a11y/no-noninteractive-tabindex */
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import styles from './ToolbarContainer.scss';
@@ -21,6 +22,7 @@ class ToolbarContainer extends Component {
   }
 
   setToolbarContainerRef = ref => (this.toolbarContainerRef = ref);
+  focus = () => this.toolbarContainerRef.focus();
 
   getToolbarPosition = () => {
     const { toolbarPosition } = this.props;
@@ -48,6 +50,7 @@ class ToolbarContainer extends Component {
     const { toolbarPosition } = this.state;
     return (
       <div
+        tabIndex={0}
         ref={this.setToolbarContainerRef}
         className={styles.container}
         style={{