+++ CHANGELOG.md
@@ -15,7 +15,8 @@
     Changes that have landed in master but are not yet released.
     Click to see more.
   </summary>
-  
+
+### :rocket: New Feature
 - `image-plugin`
   - [#2176](https://github.com/wix-incubator/rich-content/pull/2176) update image settings modal to have download and expand options using toggle buttons